+++ plugins/canvas/src/CanvasFeatureRenderer/CanvasFeatureRenderer.ts
@@ -1,6 +1,6 @@
-import { readConfObject } from '@jbrowse/core/configuration'
 import { BoxRendererType } from '@jbrowse/core/pluggableElementTypes'
-import { renderToAbstractCanvas, updateStatus } from '@jbrowse/core/util'
+
+import { doAll } from './doAll'
 
 import type { RenderArgsDeserialized } from '@jbrowse/core/pluggableElementTypes/renderers/BoxRendererType'
 
@@ -10,62 +10,19 @@ export default class CanvasFeatureRenderer extends BoxRendererType {
   async render(renderProps: RenderArgsDeserialized) {
     const features = await this.getFeatures(renderProps)
     const layout = this.createLayoutInWorker(renderProps)
-    const { statusCallback = () => {}, regions, bpPerPx, config } = renderProps
+    const { regions, bpPerPx } = renderProps
     const region = regions[0]!
     const width = Math.max(1, (region.end - region.start) / bpPerPx)
 
-    // Compute layouts for all features
-    const layoutRecords = await updateStatus(
-      'Computing feature layout',
-      statusCallback,
-      async () => {
-        const { computeLayouts } = await import('./computeLayouts')
-        return computeLayouts({
-          features,
-          bpPerPx,
-          region,
-          config,
-          layout,
-        })
-      },
-    )
+    // Render to canvas
+    const res = await doAll({
+      pluginManager: this.pluginManager,
+      renderProps,
+      layout,
+      features,
+    })
 
     const height = Math.max(1, layout.getTotalHeight())
-
-    // Fetch peptide data for CDS features
-    const peptideDataMap = await updateStatus(
-      'Fetching peptide data',
-      statusCallback,
-      async () => {
-        const { fetchPeptideData } = await import('./peptideUtils')
-        return fetchPeptideData(this.pluginManager, renderProps, features)
-      },
-    )
-
-    // Render to canvas
-    const res = await updateStatus(
-      'Rendering features',
-      statusCallback,
-      async () => {
-        const { makeImageData } = await import('./makeImageData')
-        const displayMode = readConfObject(config, 'displayMode') as string
-
-        return renderToAbstractCanvas(width, height, renderProps, ctx =>
-          makeImageData({
-            ctx,
-            layoutRecords,
-            renderArgs: {
-              ...renderProps,
-              features,
-              layout,
-              displayMode,
-              peptideDataMap,
-              colorByCDS: (renderProps as any).colorByCDS,
-            },
-          }),
-        )
-      },
-    )
 
     const result = await super.render({
       ...renderProps,
