+++ products/jbrowse-web/src/tests/MultiBigWig.test.tsx
@@ -1,13 +1,6 @@
 import { fireEvent } from '@testing-library/react'
 
-import {
-  createView,
-  doBeforeEach,
-  expectCanvasMatch,
-  hts,
-  pv,
-  setup,
-} from './util'
+import { createView, doBeforeEach, expectCanvasMatch, hts, setup } from './util'
 
 setup()
 
@@ -15,12 +8,12 @@ beforeEach(() => {
   doBeforeEach()
 })
 
-const delay = { timeout: 50000 }
+const delay = { timeout: 60000 }
 const opts = [{}, delay]
 
 test('open a multibigwig track', async () => {
-  const { view, findByTestId } = await createView()
+  const { view, findAllByTestId, findByTestId } = await createView()
   view.setNewView(5, 0)
   fireEvent.click(await findByTestId(hts('volvox_microarray_multi'), ...opts))
-  expectCanvasMatch(await findByTestId(pv('1..4000-0'), ...opts))
-}, 50000)
+  expectCanvasMatch((await findAllByTestId(/prerendered_canvas/, ...opts))[0]!)
+}, 60000)
